close #1 setup mvc folder strucutre
