+++ lib/app/app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:mobile_project/app/ui/screens/home/home_view.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const HomeView(),
+    );
+  }
+}
