+++ lib/app/ui/screens/home/home_view.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class HomeView extends StatelessWidget {
+  const HomeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Clone Spotify"),
+      ),
+    );
+  }
+}
